test/fixtures/wpt/resource-timing/shared-worker-rt-entry.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing Entry for Shared Workers</title>
<link rel="help" href="https://w3c.github.io/resource-timing/"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
</head>
<body>
<script>
    promise_test(async () => {
        const url = new URL(`resources/shared-worker.js?${token()}`, location.href).toString();
        const worker = new SharedWorker(url, 'name');
        const {data} = await new Promise(resolve => {
            worker.port.onmessage = resolve;
        });

        const timeOrigin = data;
        const entries = performance.getEntriesByName(url);
        assert_equals(entries.length, 0, "SharedWorker should not create a ResourceTiming entry");
    }, "Shared workers should not generate Resource Timing Entries");
</script>
</body>
</html>
